requirements.txt
python==3.6.9
numpy==1.16.5
pillow==6.2.0
pytorch==1.4.0
torchvision==0.5.0
scikit-learn==0.21.3
pytz==2019.3
colorama==0.4.1
tqdm==4.36.1